# api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_world_api_v1_worlds_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_world_api_v1_worlds_post",
        "type": "object"
      },
      "Body_upload_world_api_world_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_world_api_world_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OkResponse",
        "type": "object"
      },
...
      "OperationRef": {
        "description": "202 响应：去 GET /api/v1/operations/{id} 查进度。",
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "operation_id": {
            "title": "Operation Id",
            "type": "string"
          },
          "poll": {
            "default": "/api/v1/operations/{operation_id}",
            "title": "Poll",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "operation_id",
          "state",
          "kind"
        ],
        "title": "OperationRef",
        "type": "object"
      },
...
      "SwitchWorldRequest": {
        "properties": {
          "file": {
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "SwitchWorldRequest",
        "type": "object"
      },
...
      "UploadResponse": {
        "properties": {
          "file": {
            "title": "File",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "file",
          "size"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorldRef": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "file": {
            "title": "File",
            "type": "string"
          },
          "modified_at": {
            "title": "Modified At",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "file",
          "name",
          "size",
          "modified_at",
          "active"
        ],
        "title": "WorldRef",
        "type": "object"
      },
      "WorldSwitchResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          },
          "world": {
            "title": "World",
            "type": "string"
          }
        },
        "required": [
          "world",
          "message"
        ],
        "title": "WorldSwitchResponse",
        "type": "object"
      },
      "WorldUploadResponse": {
        "properties": {
          "file": {
            "title": "File",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "file",
          "size"
        ],
        "title": "WorldUploadResponse",
        "type": "object"
      },
...
      "app__schemas__v1__WorldListResponse": {
        "properties": {
          "active_world": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active World"
          },
          "backup_dir": {
            "title": "Backup Dir",
            "type": "string"
          },
          "worlds": {
            "items": {
              "$ref": "#/components/schemas/WorldRef"
            },
            "title": "Worlds",
            "type": "array"
          }
        },
        "required": [
          "worlds",
          "active_world",
          "backup_dir"
        ],
        "title": "WorldListResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Terraria Server API",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/worlds": {
      "get": {
        "operationId": "list_worlds_api_v1_worlds_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__v1__WorldListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Worlds",
        "tags": [
          "v1:worlds"
        ]
      },
      "post": {
        "operationId": "upload_world_api_v1_worlds_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_world_api_v1_worlds_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload World",
        "tags": [
          "v1:worlds"
        ]
      }
    },
    "/api/v1/worlds/{file}": {
      "delete": {
        "operationId": "delete_world_api_v1_worlds__file__delete",
        "parameters": [
          {
            "in": "path",
            "name": "file",
            "required": true,
            "schema": {
              "title": "File",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete World",
        "tags": [
          "v1:worlds"
        ]
      }
    },
    "/api/v1/worlds/{file}/activate": {
      "post": {
        "operationId": "activate_world_api_v1_worlds__file__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "file",
            "required": true,
            "schema": {
              "title": "File",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationRef"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate World",
        "tags": [
          "v1:worlds"
        ]
      }
    },
    "/api/v1/worlds/{file}/backup": {
      "post": {
        "operationId": "backup_world_api_v1_worlds__file__backup_post",
        "parameters": [
          {
            "in": "path",
            "name": "file",
            "required": true,
            "schema": {
              "title": "File",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationRef"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Backup World",
        "tags": [
          "v1:worlds"
        ]
      }
    },
...
    "/api/world/switch": {
      "post": {
        "operationId": "switch_world_api_world_switch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchWorldRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldSwitchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Switch World",
        "tags": [
          "world"
        ]
      }
    },
    "/api/world/upload": {
      "post": {
        "operationId": "upload_world_api_world_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_world_api_world_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload World",
        "tags": [
          "world"
        ]
      }
    }
  }
}